requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistics
statsmodels>=0.14.0

# Neuroimaging
nibabel>=5.0.0
nilearn>=0.10.0

# BIDS handling
pybids>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Optional: Jupyter for interactive analysis
# jupyter>=1.0.0
# ipykernel>=6.22.0

